Clean up some "robot resource leaks."
Specifically, remove robots that are no longer needed from the
Robot.robotTable so that this data structure does not grow indefinitely.
(This is not to be confused with robot.robotTable, which is filled with
static data once and never updated after that.)

When you examine the robot.robotTable during a game now, you should only see
2N robots, where N is the number of robots that the game began with.  Half of
the robots are actually playing, while the other half serve as "pristines"
for cloning purposes.  (The cloning bit is probably not necessary anymore, but
I can address that later.)

There are a small number of places where we create temporary robots that
aren't meant to play the game; I guess the best practice for now is to
immediately unregister them, though a case could be made that it ought to be
a constructor argument.