Resolve some issues I had with the Executioner.
The Executioner's main weapon, a doubled-up Mortar Cannon, had far too high a
probability of dealing the maximum possible damage.  I've replaced it with a
new weapon, the "Heavy Pulse" (not "impulse"), that has the same
carefully-chosen expected value but a more gentle damage curve.  You will
note that the maximum damage has gone *up* to 41, but this will be rare.

To summarize: Executioners and Man-O-Wars have weapons with the same EV.  The
Executioners have more of them, and better armor.  But they cost more and
have fewer hitpoints, so Man-O-Wars win more than half the time in practice.

I also made the Stormcrow's AC/20 actually have 20 ammo, since it otherwise
has to resort to that LRM peashooter too often.