Calculate the combat effectiveness more...effectively.
I'm not so sure the statistic I devised during the previous commit is as much
a measure of effectiveness as it is a measure of AI priority.  Dangerous Bots
tend to have low scores because the AI kills them early, whereas the least
harmful Bots with the highest scores are essentially allowed to get away with
murder until the endgame.

Lower numbers mean the bot tends to die first.

So I calculate three different statistics for each robot type instead: the
average robot longevity (turns taken) per match; the average originalDamage
dealt per turn, per robot, per selection screen combat point (called the "cost
effectiveness"); and the average finalDamage dealt per game, per robot (called
the "combat effectiveness.")

The Scarabs and Hermes are the most cost effective units in the game.
Obviously, the assault bots deal the most damage.  The longest-lasting units
are the heavily-armored assault bots, the Kappa, and the Stormcrow (the latter
two being the biggest jumping units.)