Make the AI more effective in hopelessly outmatched situations.
Suppose you have an enemy 'Bot that has -- I dunno, 10 armor.  Now the
expected damage that the Hermes's and and Scarab's weapons will deal against
that enemy is 0.  And in cases where the W3 (expected damage by weapons with
ammunition) and W4 (without ammunition) weights are both 0, the computer would
simply use the W4 weight since that was the last to be computed.

For the Hermes, that means firing the dual light lasers (3 + 3 damage) at the foe,
and for the Scarab that means firing a single medium laser (2d4 damage).
Neither have any hope of ever piercing through 10 points of armor, so they are
the _wrong choice_ regardless of equal weight.

In both of the above cases, the W3 weapon (the Hermes' 1 + 1d12 medium pulse)
and the Scarab's 2d7 machine gun) are the right choice.  Even though their
expected damages (7.5 and 8 respectively) cannot penetrate 10 points of armor,
their maximum damages (13 and 14) _can_.  So I instituted the concept of
an "emergency weapon" to be used when the W3 and W4 weights were both 0.

The result is that these light 'Bots now bravely attack heavily armored foes
with their strongest weapon.  Even though on average they won't deal any
damage, they'll keep trying, by Gum, until a dent is made.