Found the root cause of the "computers ganging up on me" problem.
When the computer triesd to determine who to shoot, it assigns scores to
various robots based on their dangerousness and vulnerabilities.  At the
beginning of the game, though, a lot of these scores are identical -- my
Marauder^WKraken is just as dangerous as the Machine Guild's Krakens.

In those cases, the computer previously chose the first match and shot at
that.  But since the human player tends to go first, that means that "the
first match" is usually a human Bot.

Thus I watched in awe at the beginning of the game as not one, but every
computer player systematically annihilated all of my dangerous Bots even
though they had the same ones on their own teams.

The solution right now is to gather the matches in a temporary array and
choose one at random.  A better solution for later might be to remember which
player has been annoying us the most and shoot at them.  (But that would
better be done with a global "w6" weight, methinks.)