Add support for up to 36 robots instead of 25.
I tried a few large matches yesterday and it was a real shame that the game
crashed with JavaScript errors right before a large battle.

This solution isn't perfect; I'll probably need to do this again to support up
to 49 players later.  However, keep in mind that the Bot names are not
truncated right now, so their widths *can* run over if the robot divs are too
small.  That should be an easy CSS fix, though.