Add a new robot for the first time in a while: The Ullr.
The Bullfrog is a lightly-armored Bot with heavy weaponry; the Ullr is a
heavily-armored Bot with light weaponry.  Both have the same number of
hitpoints (20), but in practice, the computer destroys the Bullfrogs as
quickly as it can and leaves the Ullrs alone until endgame.

It's not clear what niche this fills yet.  It beats Nomads in practice though
it's a nearly even fight; it trounces other light Bots; and in turn, anything with
an assault weapon or two makes quick work of it.

I also added a new weapon effect type for light pulses.  In turn, it uses two
new explosion types, so I added the sprite sheets for those.

Everything's fully attributed.