Due to advances in armor technology, M1A1 armor now rates at 3.9.
(This has nothing to do with not wanting the Ullr to have a ridiculous
2.91 *meters* of armor plating.  Nothing at all.)

I also slowed down the Ullr because that armor is _heavy_, blast it, and those
motors aren't large enough to be fast.  It's still faster than an assault Bot,
though.

Also added a minor copyedit of the Nomad description.