Start work on the EMP blast.
No credits for the images yet--I'll get to that.

The current bug has to do with the small explosions used for the
"finisher" (phase 4 of the blast): they're not centered where the blast itself
is, and I'm not sure why.  I'll fix it later.

Everything else looks good to me--I think the blast's appearance helps to lend
credence to the fact that it is currently the game's strongest weapon.

One more thing: getting the new e21 blast to loop well was not easy.  It's
sensitive to timing, so the number of frames per second and the number of
seconds you're playing the animation matter.  I've never had to deal with a
sprite like that before, but I think I get it now.